Consolidate Cache implementations into cache.go/cache_test.go
