+++ v2/cache_test.go
@@ -3,7 +3,10 @@ package pi
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"testing"
+
+	"github.com/alicebob/miniredis"
 )
 
 const (
@@ -19,9 +22,9 @@ func TestNoopCache(t *testing.T) {
 	if cache == nil {
 		t.Error("Noop cache is nil")
 	}
-	for i := 0; i < TEST_CACHE_LOOP_LIMIT; i++ {
+	for i := uint64(0); i < TEST_CACHE_LOOP_LIMIT; i++ {
 		expected := ""
-		key := fmt.Sprintf("%d", i)
+		key := strconv.FormatUint(i, 16)
 		actual, err := cache.GetValue(ctx, key)
 		if err != nil {
 			t.Errorf("GetValue returned an error: %v", err)
@@ -30,6 +33,45 @@ func TestNoopCache(t *testing.T) {
 			t.Errorf("Index %d: Expected %s received %s", i, expected, actual)
 		}
 		if err = cache.SetValue(ctx, key, "1234"); err != nil {
+			t.Errorf("Index: %d: SetValue returned an error: %v", i, err)
+
+		}
+		actual, err = cache.GetValue(ctx, key)
+		if err != nil {
+			t.Errorf("GetDigits returned an error: %v", err)
+		}
+		if actual != expected {
+			t.Errorf("Index %d: Expected %s received %s", i, expected, actual)
+		}
+
+	}
+}
+
+// The RedisCache will use a Redis-like in-memory instance to cache values. The
+// test should confirm that a value can be added to the cache and recalled
+// successfully.
+func TestRedisCache(t *testing.T) {
+	ctx := context.Background()
+	mock, err := miniredis.Run()
+	if err != nil {
+		t.Errorf("Error running miniredis: %v", err)
+	}
+	cache := NewRedisCache(ctx, mock.Addr())
+	if cache == nil {
+		t.Error("Redis cache is nil")
+	}
+	for i := uint64(0); i < TEST_CACHE_LOOP_LIMIT; i++ {
+		expected := ""
+		key := strconv.FormatUint(i, 16)
+		actual, err := cache.GetValue(ctx, key)
+		if err != nil {
+			t.Errorf("GetValue returned an error: %v", err)
+		}
+		if actual != expected {
+			t.Errorf("Index %d: Expected %s received %s", i, expected, actual)
+		}
+		expected = fmt.Sprintf("%09d", i)
+		if err = cache.SetValue(ctx, key, expected); err != nil {
 			t.Errorf("Index: %d: SetValue returned an error: %v", i, err)
 
 		}
