+++ v2/pi_test.go
@@ -2,19 +2,56 @@ package pi
 
 import (
 	"context"
+	"strconv"
 	"testing"
+
+	"github.com/alicebob/miniredis"
 )
 
-func TestPiDigit(t *testing.T) {
+func TestFractionalDigit_WithNoopCache(t *testing.T) {
 	ctx := context.Background()
-	for i := 0; i < 100; i++ {
-		expected := string(PI_DIGITS[i])
-		actual, err := PiDigit(ctx, uint64(i))
+	cache := NewNoopCache()
+	if cache == nil {
+		t.Error("Noop cache is nil")
+	}
+	SetCache(cache)
+	for i := uint64(0); i < 9*TEST_CACHE_LOOP_LIMIT; i++ {
+		expected, err := strconv.ParseUint(PI_DIGITS[i:i+1], 10, 32)
 		if err != nil {
-			t.Errorf("Error calling PiDigit: %v", err)
+			t.Errorf("Error parsing Uint: %v", err)
 		}
-		if actual != expected {
-			t.Errorf("Checking offset: %d: expected %s got %s", i, expected, actual)
+		actual, err := FractionalDigit(ctx, i)
+		if err != nil {
+			t.Errorf("Error calling FractionalDigit: %v", err)
+		}
+		if actual != uint32(expected) {
+			t.Errorf("Checking offset: %d: expected %d got %d", i, expected, actual)
+		}
+	}
+}
+
+func TestFractionalDigit_WithRedisCache(t *testing.T) {
+	ctx := context.Background()
+	mock, err := miniredis.Run()
+	if err != nil {
+		t.Errorf("Error running miniredis: %v", err)
+	}
+	cache := NewRedisCache(ctx, mock.Addr())
+	if cache == nil {
+		t.Error("Redis cache is nil")
+	}
+	SetCache(cache)
+	for i := uint64(0); i < 9*TEST_CACHE_LOOP_LIMIT; i++ {
+		expected, err := strconv.ParseUint(PI_DIGITS[i:i+1], 10, 32)
+		if err != nil {
+			t.Errorf("Error parsing Uint: %v", err)
+		}
+		actual, err := FractionalDigit(ctx, i)
+		if err != nil {
+			t.Errorf("Error calling FractionalDigit: %v", err)
+		}
+		if actual != uint32(expected) {
+			t.Errorf("Checking offset: %d: expected %d got %d", i, expected, actual)
 		}
 	}
 }
